requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
openai==1.54.3
numpy==1.26.3
scikit-learn==1.4.0
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
httpx==0.27.0